+++ clang/lib/Format/Format.cpp
@@ -207,12 +207,12 @@ template <> struct ScalarEnumerationTraits<FormatStyle::BracketAlignmentStyle> {
   static void enumeration(IO &IO, FormatStyle::BracketAlignmentStyle &Value) {
     IO.enumCase(Value, "Align", FormatStyle::BAS_Align);
     IO.enumCase(Value, "DontAlign", FormatStyle::BAS_DontAlign);
-    IO.enumCase(Value, "AlwaysBreak", FormatStyle::BAS_AlwaysBreak);
-    IO.enumCase(Value, "BlockIndent", FormatStyle::BAS_BlockIndent);
 
     // For backward compatibility.
     IO.enumCase(Value, "true", FormatStyle::BAS_Align);
     IO.enumCase(Value, "false", FormatStyle::BAS_DontAlign);
+    IO.enumCase(Value, "AlwaysBreak", FormatStyle::BAS_AlwaysBreak);
+    IO.enumCase(Value, "BlockIndent", FormatStyle::BAS_BlockIndent);
   }
 };
 
@@ -1250,6 +1250,37 @@ template <> struct MappingTraits<FormatStyle> {
                    Style.WhitespaceSensitiveMacros);
     IO.mapOptional("WrapNamespaceBodyWithEmptyLines",
                    Style.WrapNamespaceBodyWithEmptyLines);
+
+    // If AlwaysBreak or BlockIndent were specified but individual
+    // options for BreakAfterOpenBracket* (CloseAfterOpenBracket*),
+    // initialize the latter to preserve backwards compatibility.
+    if (Style.AlignAfterOpenBracket == FormatStyle::BAS_AlwaysBreak) {
+      if (!Style.BreakAfterOpenBracketBracedList &&
+          !Style.BreakAfterOpenBracketFunction &&
+          !Style.BreakAfterOpenBracketIf && !Style.BreakAfterOpenBracketLoop &&
+          !Style.BreakAfterOpenBracketSwitch) {
+        Style.BreakAfterOpenBracketBracedList = true;
+        Style.BreakAfterOpenBracketFunction = true;
+        Style.BreakAfterOpenBracketIf = true;
+      }
+    } else if (Style.AlignAfterOpenBracket == FormatStyle::BAS_BlockIndent) {
+      if (!Style.BreakAfterOpenBracketBracedList &&
+          !Style.BreakAfterOpenBracketFunction &&
+          !Style.BreakAfterOpenBracketIf && !Style.BreakAfterOpenBracketLoop &&
+          !Style.BreakAfterOpenBracketSwitch &&
+          !Style.BreakBeforeCloseBracketBracedList &&
+          !Style.BreakBeforeCloseBracketFunction &&
+          !Style.BreakBeforeCloseBracketIf &&
+          !Style.BreakBeforeCloseBracketLoop &&
+          !Style.BreakBeforeCloseBracketSwitch) {
+        Style.BreakAfterOpenBracketBracedList = true;
+        Style.BreakAfterOpenBracketFunction = true;
+        Style.BreakAfterOpenBracketIf = true;
+        Style.BreakBeforeCloseBracketBracedList = true;
+        Style.BreakBeforeCloseBracketFunction = true;
+        Style.BreakBeforeCloseBracketIf = true;
+      }
+    }
 
     // If AlwaysBreakAfterDefinitionReturnType was specified but
     // BreakAfterReturnType was not, initialize the latter from the former for